tabs dashboard

+++ admin/models/product.php
@@ -208,28 +208,27 @@ function deleteProduct() {
 
 function applyProd() {
     if (isset($_POST['apply'])) {
-        global $conn;
-        $checkboxes = $_POST['checkBoxArr'];
-        foreach ($checkboxes as $checkbox) {
-            $option = $_POST['option'];
-            
-            switch ($option) {
-                case "delete":
+        if (!empty($_POST['checkBoxArr'])) {
+            global $conn;
+            $checkboxes = $_POST['checkBoxArr'];
+            foreach ($checkboxes as $checkbox) {
+                $option = $_POST['option'];
+                
+                switch ($option) {
+                    case "delete":
+                        $sql = "DELETE FROM products WHERE id = $checkbox";
+                        $statement = $conn->query($sql);
+                        if ($statement -> execute()) {
+                            header("location: index.php?act=listProd");
+                        }
+                        break;
 
-                    $sql = "DELETE FROM products WHERE id = $checkbox";
-                    $statement = $conn->query($sql);
-                    if ($statement -> execute()) {
-                        header("location: index.php?act=listProd");
-                    }
-                    break;
-
-                default:
-
-                    echo "<script>Swal.fire('Bạn phải chọn chức năng')</script>";
-                    break;
+                    default:
+                        echo "<script>Swal.fire('Bạn phải chọn chức năng')</script>";
+                        break;
+                }
 
             }
-
         }
     }
 }
